requirements.txt
# ─── Core ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.1
schedule==1.2.1

# ─── PDF / NLP ────────────────────────────────────────────────────────────────
pdfplumber==0.11.0
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# ─── Database ─────────────────────────────────────────────────────────────────
mysql-connector-python==8.3.0
SQLAlchemy==2.0.29

# ─── Vector Store ─────────────────────────────────────────────────────────────
chromadb==0.5.0
sentence-transformers==2.7.0

# ─── Web / Research ───────────────────────────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.43.0

# ─── GitHub Integration ───────────────────────────────────────────────────────
PyGithub==2.3.0

# ─── AI / LLM — Google Gemini (FREE tier available) ──────────────────────────
google-generativeai==0.7.2

# ─── Notifications ────────────────────────────────────────────────────────────
discord.py==2.3.2

# ─── Dashboard ────────────────────────────────────────────────────────────────
flask==3.0.3
flask-sqlalchemy==3.1.1

# ─── Static Analysis ──────────────────────────────────────────────────────────
pylint==3.1.0
bandit==1.7.8
lizard==1.17.10

# ─── Utilities ────────────────────────────────────────────────────────────────
watchdog==4.0.0
python-crontab==3.0.0
loguru==0.7.2
rich==13.7.1
